Fix direct CSR
By moving from environment.ts to config.yml we made it so the environment is _not_ up to date with the server configuration when main.js is first loaded.
Because of this the app behaved as if CSR always happened _after_ SSR, effectively breaking direct CSR.

Here the "criterion" for SSR/non-SSR HTML is changed from the related configuration property to the presence of Angular Universal transfer state.

This means we can correctly determine when to bootstrap the app for direct CSR, and it's' now "safe" to just send index.html by itself.