ESLint: enforce prefer-const where possible
https://eslint.org/docs/latest/rules/prefer-const

In our tests we use this pattern a lot:
```
let something;

beforeEach(() => {
  something = ...;
});
```

This is not valid ~ prefer-const, but can be "let through" with some configuration.
However, this would make the rule too loose for the non-test codebase.

Therefore, we enable it overall but turn it off for tests only -- the `*.spec.ts` overrides apply to test files only and take precedence over `*.ts`