107873: Shorten paginationId
I chose 2 characters instead of one in case a 'SUCCEEDED' process status
is added in the future. 'SUCCEEDED' could encompass both 'COMPLETED' and
'FAILED' processes, but it's paginationId would clash with 'SCHEDULED'
(both start with 's'). By choosing the first 2 characters of the status,
this clash does not occur (ids would be 'sc' and 'su').