Use npm instead of yarn in Dockerfile
There is a minor difference in syntax when starting the application
with `npm run serve`. We use `--` to make sure npm doesn't try to
parse the `--host 0.0.0.0` as an option, allowing it instead to be
read by Angular CLI (ng).

Also, there is no direct npm equivalent to yarn's --network-timeout
option so I will remove it. If we need something similar later we
might be able to use these config options:

    npm config set fetch-retry-mintimeout 300000
    npm config set fetch-retry-maxtimeout 300000