119176: Make table horizontally scrollable
For most screen sizes, the ResponsiveTableSizes is enough to resize the
table columns. On very small screens, or when zoomed in a lot, even the
smallest column sizes are too big. To make it possible to view the
rest of the content even in these situations, the ability to scroll
horizontally is added.