Add undeclared body-parser dependency
We explicitly use this in server.ts. It has always been available
because it is a dependency of express, and therefore a transitive
dependency of ours. Because we are using it directly in *our* code
we should declare the dependency.

I am using version ^1.20.3 because that is the one used by the
version of Express we are using.